+++ src/atlas/interpolation/method/knn/KNearestNeighbours.cc
@@ -97,7 +97,7 @@ void KNearestNeighbours::setup( const FunctionSpace& source, const FunctionSpace
             }
 
             // find the closest input points to the output point
-            PointIndex3::Point p{coords( ip, 0 ), coords( ip, 1 ), coords( ip, 2 )};
+            PointIndex3::Point p{coords( ip, (size_t)0 ), coords( ip, (size_t)1 ), coords( ip, (size_t)2 )};
             PointIndex3::NodeList nn = pTree_->kNearestNeighbours( p, k_ );
 
             // calculate weights (individual and total, to normalise) using distance
